requirements.txt
setuptools==49.6.0
wheel==0.35.1

google-cloud-secret-manager==1.0.0
google-cloud-storage==1.30.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
pyyaml==5.3.1